Change double quotes to single quotes
"." is a string literal, and so is treated as the char '.' plus the null char '\0'. Single quotes reduces the necessary memory for this literal to only one char instead. 
A string literal as the one actually present may require the use of the method "write(const char *str)", so there could be also a performance overhead. 
Another reason to change quotes style is for consistency with line 235.